% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split}
\alias{split}
\title{reads in xy data frame of coordinates and a distance value to split line on}
\usage{
split(xydf, dist)
}
\arguments{
\item{xydf}{dataframe with x and y coords}

\item{dist}{distance in units of coords}
}
\value{
split version of xydf
}
\description{
to be filled
}
