% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitPoints.R
\name{splitPoints}
\alias{splitPoints}
\title{Split Points}
\usage{
splitPoints(spobj, dist, start = T, simplify = F)
}
\arguments{
\item{spobj}{sp object}

\item{dist}{distance in units of coords}

\item{start}{start from the start (T) or the end (F)}

\item{simplify}{return simple version x/y points if T else, SpatialPointsDataFrame}
}
\value{
points split in format per simplify
}
\description{
to be filled
}
