% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitLines.R
\name{splitLines}
\alias{splitLines}
\title{Split Lines}
\usage{
splitLines(spobj, dist, start = T)
}
\arguments{
\item{spobj}{sp object}

\item{dist}{distance in units of coords}

\item{start}{start from the start (T) or the end (F)}
}
\value{
spatialLinesDataFrame
}
\description{
to be filled
}
